S3a: fire FIELD -> int32 Q16.16 + the Python ignition twin (representation-only)
The THIRD and FINAL field migration of the fixed-point arc, sub-step S3a
(docs/s3_fixed_point_plan.md). Flips the `fire` storage to int32 Q16.16 and makes
BOTH fire-write paths integer-consistent, WITHOUT yet converting the C++ logistic
math — the C++ FireSimulation::step keeps running in FLOAT behind a temporary
internal bridge (the S2 internal-bridge discipline), so the FEEL is unchanged this
commit. The math goes integer in S3b; the bridges are deleted in S3c.

Field + helper
- gmap.fire -> int32 Q16.16 (gamemap.py). New src/simulation/fire_fixed.py mirrors
  gas_fixed.py (quantize/dequantize boundary helpers, FP_ONE=65536; GPU width int16
  Q1.15 recorded in the format tag, CPU ships int32 — same decide-once as smoke/gas).

The Python ignition twin (combat.apply_temperature_ignition)
- O2 gate: replaced the float dequantize + float mean with an INTEGER int64
  neighbour-sum + round-half-away-from-zero mean (== fixed_point.h mean_round) on
  the int32 atmosphere, over the OPEN (~solid & ~is_vacuum) 4-neighbour mask — the
  EXACT C++ fire P-gate mask (the pre-S3a Python mask `~solid` wrongly INCLUDED
  vacuum neighbours, lowering the mean; excluding vacuum is what makes the two O2
  predicates agree — the design-note invariant). o2_threshold quantized once.
- fire write: ignition_seed -> quantize_scalar; np.maximum is now an integer max.

gmap.fire reader audit (a dequantize at every boundary)
- C++ bridge: physics_engine.step_tail dequantizes the int32 fire into a reused
  fire_f_ float scratch, runs the still-float fire.step, re-quantizes back
  (round-to-nearest); the step_tail binding + signature take int32 fire now.
- cast_fire_heat: the `>0` burning mask stays integer; the per-tile intensity that
  feeds the heat-ray range/intensity params is dequantized (local/cosmetic + the
  heat payload, float-OK).
- renderer (game_renderer.upload_state), recorder, the field_edit "fire" policy
  (new _combine_fire, dequantize->combine->requantize like gas), and the
  input_handler debug seed all get a dequantize/quantize boundary.

Gates
- P1 self-match tol=0.0 (field_ab_harness: fire now int32 in SIM_FIELDS, bit-
  identical run-to-run; the harness fire seed is quantized so the fire feedback +
  its smoke plume stay live — a raw 0.8 would store 0 counts).
- O2-invariant (tests/test_fire_o2_invariant.py, the headline gate): the Python
  integer O2 gate and the C++ fire float P gate return the identical boolean across
  a homogeneous + heterogeneous atmosphere sweep; the ONLY disagreements are EXACT
  threshold ties (integer mean == quantize(0.60)), where the integer gate is the
  correct/deterministic one and S3b's C++ mean_round bit-matches it.
- Ignition path exercised; existing fire/ignition/field-edit tests updated to the
  int32 representation. Full suite 360 green; test_main_smoke.py --auto exit 0.
- no-float ratchet: physics_engine.cpp baseline 65/31 -> 68/32 (the temporary fire
  bridge's comment lines + the re-quantize (double) fold). fire_simulation.cpp
  UNCHANGED (its logistic is float until S3b).

+++ cpp/src/bindings.cpp
@@ -553,7 +553,7 @@ PYBIND11_MODULE(breach_physics, m) {
                              py::array_t<int32_t> wave_p,        // S2a: Q16.16 int32
                              py::array_t<bool>  solid,
                              // fire group
-                             py::array_t<float> fire,
+                             py::array_t<int32_t> fire,          // S3a: Q16.16 int32 (fire field bridge inside)
                              py::array_t<int32_t> atmosphere,   // S2c: Q16.16 int32 (fire bridge inside)
                              py::array_t<int32_t> smoke,         // S2b: Q16.16 int32
                              py::array_t<float> wall_hp,
